refactor: reorganize validation and auth logic
